move into PD from helper functions
